ai/requirements.txt
fastapi~=0.115.12
uvicorn~=0.34.3
ultralytics~=8.3.153
opencv-python~=4.11.0.86
numpy~=2.2.6
Pillow~=10.3.0
python-multipart==0.0.6
requests==2.31.0
pydantic~=2.11.5
paddleocr~=3.4.0